*: move bench channel to bench directory.
