+++ benches/channel/mod.rs
@@ -1,0 +1,1 @@
+mod bench_channel;
